requirements.txt
pystrum
torchio
